+++ BirdsEye/sim/stubs/module_stubs.cpp
@@ -134,6 +134,7 @@ void NEOPIXEL_LOOP() {}
 void NEOPIXEL_SLEEP() {}
 void NEOPIXEL_WAKE() {}
 void neopixelNotifyPurpleSector() {}
+void neopixelNotifyPurpleLap() {}
 
 // ---- profiling.ino surface ----